+++ packages/blockchain-api/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 0.11.20
+
+### Patch Changes
+
+- Updated dependencies [[`77df26b`](https://github.com/helium/helium-program-library/commit/77df26b20ce9922b11f6b6e36b9f45b1a723e8bc)]:
+  - @helium/blockchain-api@0.14.0
+
+## 0.11.19
+
+### Patch Changes
+
+- Updated dependencies [[`a5d7e07`](https://github.com/helium/helium-program-library/commit/a5d7e073f3da1ab87816c982ec723c7e2158a5ac), [`a5d7e07`](https://github.com/helium/helium-program-library/commit/a5d7e073f3da1ab87816c982ec723c7e2158a5ac)]:
+  - @helium/voter-stake-registry-sdk@0.12.0
+  - @helium/blockchain-api@0.13.0
+  - @helium/helium-sub-daos-sdk@0.11.17
+  - @helium/hpl-crons-sdk@0.11.17
+
 ## 0.11.18
 
 ### Patch Changes